Moved methods from page model to page service
